Players can place items on the chopping board, and a change to input
Now players can take food they have picked up and place it on either chopping board, the rest of the chopping functionality must be refined however. An important change was made to player input, tracking it in Update instead of OnTriggerStay (which runs more like a FixedUpdate and will not always catch player input).